backend/pyproject.toml
[project]
name = "spendly-backend"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.30",
  "psycopg2-binary>=2.9",
  "alembic>=1.13",
  "pydantic>=2.0",
  "pydantic-settings>=2.0",
  "pdfplumber>=0.11",
  "camelot-py[cv]>=0.11",
  "pandas>=2.0",
  "qdrant-client>=1.9",
  "nats-py>=2.7",
  "redis[hiredis]>=5.0",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "python-jose[cryptography]>=3.3",
  "pyyaml>=6.0",
  "rapidfuzz>=3.0",
  "python-dateutil>=2.9",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
